+++ lib/verboten_keys.rb
@@ -1,15 +1,20 @@
 # frozen_string_literal: true
 
-require 'rack'
+require "rack"
 
-require_relative 'verboten_keys/version'
-require_relative 'verboten_keys/errors'
-require_relative 'verboten_keys/configuration'
-require_relative 'verboten_keys/filterer'
-require_relative 'verboten_keys/middleware'
-require_relative 'verboten_keys/railtie' if defined?(::Rails)
+require_relative "verboten_keys/version"
+require_relative "verboten_keys/errors"
+require_relative "verboten_keys/configuration"
+require_relative "verboten_keys/filterer"
+require_relative "verboten_keys/middleware"
+require_relative "verboten_keys/railtie" if defined?(Rails)
 
 module VerbotenKeys
+  class ForbiddenKeyError < StandardError
+    def initialize(key)
+      super("Forbidden key '#{key}' found in response")
+    end
+  end
   class << self
     attr_accessor :configuration