+++ lib/verboten_keys/railtie.rb
@@ -2,7 +2,7 @@
 
 module VerbotenKeys
   class Railtie < ::Rails::Railtie
-    initializer 'verboten_keys.middleware' do |app|
+    initializer "verboten_keys.middleware" do |app|
       app.middleware.use VerbotenKeys::Middleware
     end
   end